Re-organize + Refactor Code
- Remove duplicate features
- Re-organize code into new packages
- Made MainActivity.kt the launcher activity
