+++ app/src/main/AndroidManifest.xml
@@ -13,23 +13,17 @@
         android:theme="@style/Theme.AlcMMSystemmentor"
         tools:targetApi="31">
         <activity
-            android:name=".ui.TaskActivity"
-            android:exported="true"
-            android:label="@string/title_activity_task">
-            <intent-filter>
-                <action android:name="android.intent.action.MAIN" />
-
-                <category android:name="android.intent.category.LAUNCHER" />
-            </intent-filter>
-            <meta-data
-                android:name="android.app.lib_name"
-                android:value="" />
-        </activity>
-        <activity
             android:name=".MainActivity"
             android:exported="true">
+            <intent-filter>
+                <action android:name="android.intent.action.MAIN" />
+                <category android:name="android.intent.category.LAUNCHER" />
+            </intent-filter>
         </activity>
-
+        <activity
+            android:name=".ui.tasks.screens.TaskActivity"
+            android:label="@string/title_activity_task">
+        </activity>
         <meta-data
             android:name="preloaded_fonts"
             android:resource="@array/preloaded_fonts" />
