--- app/src/main/java/com/peculiaruc/alc_mmsystem_mentor/adapters/CourseData.kt
@@ -1,8 +1,0 @@
-package com.peculiaruc.alc_mmsystem_mentor.adapters
-
-data class CourseInfo(val course: String, val title: String, val image: Int, val image2: Int) {
-}
-
-data class ItemTasks(val title: String, val image: Int)
-
-data class ItemPrograms(val title: String, val image: Int)
