+++ app/src/main/java/com/peculiaruc/alc_mmsystem_mentor/data/local/database/models/Course.kt
@@ -1,0 +1,8 @@
+package com.peculiaruc.alc_mmsystem_mentor.data.local.database.models
+
+data class Course(
+    val course: String,
+    val title: String,
+    val image: Int,
+    val image2: Int
+)
