+++ app/src/main/java/com/peculiaruc/alc_mmsystem_mentor/data/local/database/models/Program.kt
@@ -1,0 +1,6 @@
+package com.peculiaruc.alc_mmsystem_mentor.data.local.database.models
+
+data class Program(
+    val title: String,
+    val image: Int
+)
