+++ app/src/main/java/com/peculiaruc/alc_mmsystem_mentor/data/local/database/models/ReportTask.kt
@@ -1,0 +1,5 @@
+package com.peculiaruc.alc_mmsystem_mentor.data.local.database.models
+
+data class ReportTask(
+    val title: String,
+    val image: Int)
