+++ app/src/main/java/com/peculiaruc/alc_mmsystem_mentor/data/local/database/models/Task.kt
@@ -1,10 +1,10 @@
-package com.peculiaruc.alc_mmsystem_mentor.data.model
+package com.peculiaruc.alc_mmsystem_mentor.data.local.database.models
 
-data class MentorTasks(
-    val id:Int,
+data class Task(
+    val id: Int,
     var title: String,
     var desc: String,
     val assigned: Boolean,
     val completed: Boolean,
     val assignee: Int
-    )
+)
