--- app/src/main/java/com/peculiaruc/alc_mmsystem_mentor/mentortasks_ui/MentorsTasksDetailsFragment.kt
@@ -1,31 +1,0 @@
-package com.peculiaruc.alc_mmsystem_mentor.mentortasks_ui
-
-import android.os.Bundle
-import androidx.fragment.app.Fragment
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import com.peculiaruc.alc_mmsystem_mentor.databinding.FragmentMentorsTasksDetailsBinding
-
-class MentorsTasksDetailsFragment : Fragment() {
-
-    private var _binding: FragmentMentorsTasksDetailsBinding? = null
-
-    private val binding get() = _binding!!
-
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        _binding = FragmentMentorsTasksDetailsBinding.inflate(inflater, container, false)
-        val view = binding.root
-
-        return view
-    }
-
-    override fun onDestroyView() {
-        super.onDestroyView()
-        _binding = null
-    }
-}
