--- app/src/main/java/com/peculiaruc/alc_mmsystem_mentor/mentortasks_ui/MentorsTasksDetailsSecondFragment.kt
@@ -1,29 +1,0 @@
-package com.peculiaruc.alc_mmsystem_mentor.mentortasks_ui
-
-import android.os.Bundle
-import androidx.fragment.app.Fragment
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import com.peculiaruc.alc_mmsystem_mentor.databinding.FragmentMentorsTasksDetailsSecondBinding
-
-class MentorsTasksDetailsSecondFragment : Fragment() {
-
-    private var _binding: FragmentMentorsTasksDetailsSecondBinding? = null
-    private val binding get() = _binding!!
-
-
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        // Inflate the layout for this fragment
-        _binding = FragmentMentorsTasksDetailsSecondBinding.inflate(inflater, container, false)
-        return binding.root
-    }
-
-    override fun onDestroyView() {
-        super.onDestroyView()
-        _binding = null
-    }
-}
