+++ app/src/main/java/com/peculiaruc/alc_mmsystem_mentor/ui/reports/adapters/CourseRecyclerAdapter.kt
@@ -1,4 +1,4 @@
-package com.peculiaruc.alc_mmsystem_mentor.adapters
+package com.peculiaruc.alc_mmsystem_mentor.ui.reports.adapters
 
 import android.view.LayoutInflater
 import android.view.View
@@ -7,8 +7,9 @@ import android.widget.ImageView
 import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
 import com.peculiaruc.alc_mmsystem_mentor.R
+import com.peculiaruc.alc_mmsystem_mentor.data.local.database.models.Course
 
-class CourseRecyclerAdapter(private val courses: List<CourseInfo>) :
+class CourseRecyclerAdapter(private val courses: List<Course>) :
     RecyclerView.Adapter<CourseRecyclerAdapter.ViewHolder>() {
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewHolder {
