+++ app/src/main/java/com/peculiaruc/alc_mmsystem_mentor/ui/reports/screens/ReportDownloadFragment.kt
@@ -1,4 +1,4 @@
-package com.peculiaruc.alc_mmsystem_mentor.mentortasks_ui
+package com.peculiaruc.alc_mmsystem_mentor.ui.reports.screens
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -6,13 +6,11 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import com.peculiaruc.alc_mmsystem_mentor.R
-import com.peculiaruc.alc_mmsystem_mentor.databinding.FragmentReportSubmittedBinding
-import kotlinx.android.synthetic.main.fragment_report_submitted.*
+import com.peculiaruc.alc_mmsystem_mentor.databinding.FragmentReportDownloadBinding
+import kotlinx.android.synthetic.main.fragment_report_download.*
 
-
-class ReportSubmittedFragment : Fragment() {
-
-    private var _binding: FragmentReportSubmittedBinding? = null
+class ReportDownloadFragment : Fragment() {
+    private var _binding: FragmentReportDownloadBinding? = null
 
     private val binding get() = _binding!!
 
@@ -22,14 +20,14 @@ class ReportSubmittedFragment : Fragment() {
         savedInstanceState: Bundle?
     ): View {
         // Retrieve and inflate the layout for this fragment
-        _binding = FragmentReportSubmittedBinding.inflate(inflater, container, false)
+        _binding = FragmentReportDownloadBinding.inflate(inflater, container, false)
         return binding.root
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
 
-        btn_submitted_field.text = getString(R.string.done)
+        btn_download_field.text = getString(R.string.done)
     }
