+++ app/src/main/java/com/peculiaruc/alc_mmsystem_mentor/ui/reports/screens/ReportSubmittedFragment.kt
@@ -1,4 +1,4 @@
-package com.peculiaruc.alc_mmsystem_mentor.mentortasks_ui
+package com.peculiaruc.alc_mmsystem_mentor.ui.reports.screens
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -6,11 +6,13 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import com.peculiaruc.alc_mmsystem_mentor.R
-import com.peculiaruc.alc_mmsystem_mentor.databinding.FragmentReportDownloadBinding
-import kotlinx.android.synthetic.main.fragment_report_download.*
+import com.peculiaruc.alc_mmsystem_mentor.databinding.FragmentReportSubmittedBinding
+import kotlinx.android.synthetic.main.fragment_report_submitted.*
 
-class ReportDownloadFragment : Fragment() {
-    private var _binding: FragmentReportDownloadBinding? = null
+
+class ReportSubmittedFragment : Fragment() {
+
+    private var _binding: FragmentReportSubmittedBinding? = null
 
     private val binding get() = _binding!!
 
@@ -20,14 +22,14 @@ class ReportDownloadFragment : Fragment() {
         savedInstanceState: Bundle?
     ): View {
         // Retrieve and inflate the layout for this fragment
-        _binding = FragmentReportDownloadBinding.inflate(inflater, container, false)
+        _binding = FragmentReportSubmittedBinding.inflate(inflater, container, false)
         return binding.root
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
 
-        btn_download_field.text = getString(R.string.done)
+        btn_submitted_field.text = getString(R.string.done)
     }
