+++ app/src/main/java/com/peculiaruc/alc_mmsystem_mentor/ui/tasks/screens/TaskObjectFragment.kt
@@ -1,35 +1,25 @@
-package com.peculiaruc.alc_mmsystem_mentor.ui
+package com.peculiaruc.alc_mmsystem_mentor.ui.tasks.screens
 
 import android.os.Bundle
-import android.view.Menu
+import android.view.LayoutInflater
 import android.view.View
-import android.widget.ImageView
-import android.widget.Toast
-import androidx.appcompat.app.AppCompatActivity
-import androidx.appcompat.widget.SearchView
-import androidx.appcompat.widget.Toolbar
-import androidx.navigation.findNavController
-import androidx.navigation.ui.AppBarConfiguration
-import androidx.navigation.ui.navigateUp
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import com.peculiaruc.alc_mmsystem_mentor.R
-import com.peculiaruc.alc_mmsystem_mentor.data.model.MentorTasks
-import com.peculiaruc.alc_mmsystem_mentor.databinding.ActivityTaskBinding
-import com.peculiaruc.alc_mmsystem_mentor.ui.adapter.TaskListAdapter
-import java.util.*
+import com.peculiaruc.alc_mmsystem_mentor.data.local.database.models.Task
+import com.peculiaruc.alc_mmsystem_mentor.ui.tasks.adapters.TaskListAdapter
 
-class TaskActivity : AppCompatActivity() {
+class TaskObjectFragment(position: Int) : Fragment() {
 
-    private lateinit var appBarConfiguration: AppBarConfiguration
-    private lateinit var binding: ActivityTaskBinding
-
+    val pos: Int = position
     private lateinit var recyclerView: RecyclerView
     private lateinit var adapter: TaskListAdapter
-    var searchView: SearchView? = null
 
-    var tasks: ArrayList<MentorTasks> = arrayListOf()
+    var tasks: ArrayList<Task> = arrayListOf()
 
-    private val taskOne = MentorTasks(
+    private val taskOne = Task(
         1,
         "Write Documentation for Auth",
         "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Ut et massa mi. " +
@@ -40,7 +30,7 @@ class TaskActivity : AppCompatActivity() {
         false,
         0
     )
-    private val taskTwo = MentorTasks(
+    private val taskTwo = Task(
         2,
         "Implement Dependency Injection",
         "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Ut et massa mi. " +
@@ -50,7 +40,7 @@ class TaskActivity : AppCompatActivity() {
         false,
         3
     )
-    private val taskThree = MentorTasks(
+    private val taskThree = Task(
         3,
         "Fetch API endpoint for all tasks",
         "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Ut et massa mi. \" +\n" +
@@ -60,7 +50,7 @@ class TaskActivity : AppCompatActivity() {
         true,
         4
     )
-    private val taskFour = MentorTasks(
+    private val taskFour = Task(
         4,
         "Implement local caching",
         "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Ut et massa mi. \" +\n" +
@@ -70,7 +60,7 @@ class TaskActivity : AppCompatActivity() {
         false,
         3
     )
-    private val taskFive = MentorTasks(
+    private val taskFive = Task(
         5,
         "Create Database",
         "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Ut et massa mi. \" +\n" +
@@ -80,7 +70,7 @@ class TaskActivity : AppCompatActivity() {
         false,
         5
     )
-    private val taskSix = MentorTasks(
+    private val taskSix = Task(
         6,
         "Implement Navigation Graph",
         "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Ut et massa mi. \" +\n" +
@@ -90,7 +80,7 @@ class TaskActivity : AppCompatActivity() {
         true,
         5
     )
-    private val taskSeven = MentorTasks(
+    private val taskSeven = Task(
         7,
         "Liaise with Backend on the Settings endpoints",
         "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Ut et massa mi. \" +\n" +
@@ -100,7 +90,7 @@ class TaskActivity : AppCompatActivity() {
         false,
         1
     )
-    private val taskEight = MentorTasks(
+    private val taskEight = Task(
         8,
         "Implement Firestore caching",
         "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Ut et massa mi. \" +\n" +
@@ -110,7 +100,7 @@ class TaskActivity : AppCompatActivity() {
         false,
         3
     )
-    private val taskNine = MentorTasks(
+    private val taskNine = Task(
         9,
         "Implement UI for Chat function",
         "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Ut et massa mi. \" +\n" +
@@ -120,7 +110,7 @@ class TaskActivity : AppCompatActivity() {
         false,
         3
     )
-    private val taskTen = MentorTasks(
+    private val taskTen = Task(
         10,
         "Implement Internationalization",
         "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Ut et massa mi. \" +\n" +
@@ -131,87 +121,56 @@ class TaskActivity : AppCompatActivity() {
         4
     )
 
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        val toolbar = findViewById<Toolbar>(R.id.toolbar2)
-        setSupportActionBar(toolbar)
-
-        binding = ActivityTaskBinding.inflate(layoutInflater)
-        setContentView(binding.root)
-
-        val navController = findNavController(R.id.nav_host_fragment_content_task)
-        appBarConfiguration = AppBarConfiguration(navController.graph)
-
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View {
+        val view = inflater.inflate(R.layout.fragment_task_object, container, false)
+        recyclerView = view.findViewById<RecyclerView>(R.id.recycle)
+        adapter = TaskListAdapter()
+        recyclerView.adapter = adapter
+        recyclerView.layoutManager = LinearLayoutManager(requireContext())
+        return view
     }
 
-    override fun onSupportNavigateUp(): Boolean {
-        val navController = findNavController(R.id.nav_host_fragment_content_task)
-        return navController.navigateUp(appBarConfiguration)
-                || super.onSupportNavigateUp()
-    }
-
-    // calling on create option menu
-    // layout to inflate our menu file.
-    override fun onCreateOptionsMenu(menu: Menu): Boolean {
-        // below line is to get our inflater
-        val inflater = menuInflater
-
-        // inside inflater we are inflating our menu file.
-        inflater.inflate(R.menu.search_menu, menu)
-
-        // below line is to get our menu item.
-        val searchItem = menu.findItem(R.id.actionSearch)
-
-        // getting search view of our item.
-        val searchView: SearchView = searchItem.actionView as SearchView
-
-        searchView.setOnClickListener {
-            val backButton = findViewById<ImageView?>(R.id.imageView)
-            backButton.visibility = View.GONE
-            val title: ImageView = findViewById<ImageView?>(R.id.imageView2)
-            title.visibility = View.GONE
-            backButton.visibility = View.GONE
-        }
-
-        // below line is to call set on query text listener method.
-        searchView.setOnQueryTextListener(object : SearchView.OnQueryTextListener {
-            override fun onQueryTextSubmit(query: String): Boolean {
-                return false
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        when (pos) {
+            0 -> {
+                tasks.add(taskOne)
+                tasks.add(taskTwo)
+                tasks.add(taskThree)
+                tasks.add(taskFour)
+                tasks.add(taskFive)
+                tasks.add(taskSix)
+                tasks.add(taskSeven)
+                tasks.add(taskEight)
+                tasks.add(taskNine)
+                tasks.add(taskTen)
+                adapter.submitList(tasks)
             }
-
-            override fun onQueryTextChange(newText: String): Boolean {
-                // inside on query text change method we are
-                // calling a method to filter our recycler view.
-                filter(newText)
-                return false
+            1 -> {
+                tasks.add(taskTwo)
+                tasks.add(taskFour)
+                tasks.add(taskEight)
+                tasks.add(taskNine)
+                adapter.submitList(tasks)
             }
-        })
-        return true
-    }
-
-    private fun filter(text: String) {
-        // creating a new array list to filter our data.
-        val filteredlist = ArrayList<MentorTasks>()
-
-        adapter.submitList(tasks)
-        // running a for loop to compare elements.
-        for (item in tasks) {
-            // checking if the entered string matched with any item of our recycler view.
-            if (item.title.lowercase(Locale.getDefault()).contains(text.lowercase(Locale.getDefault()))) {
-                // if the item is matched we are
-                // adding it to our filtered list.
-                filteredlist.add(item)
+            2 -> {
+                tasks.add(taskThree)
+                tasks.add(taskSix)
+                tasks.add(taskSeven)
+                tasks.add(taskEight)
+                tasks.add(taskTen)
+                adapter.submitList(tasks)
             }
-        }
-        if (filteredlist.isEmpty()) {
-            // if no item is added in filtered list we are
-            // displaying a toast message as no data found.
-            Toast.makeText(this, "No Data Found..", Toast.LENGTH_SHORT).show()
-        } else {
-            // at last we are passing that filtered
-            // list to our adapter class.
-            adapter.filterList(filteredlist)
+            3 -> {
+                tasks.add(taskTwo)
+                tasks.add(taskFour)
+                tasks.add(taskEight)
+                tasks.add(taskNine)
+                adapter.submitList(tasks)
+            }
         }
     }
 }
