+++ app/src/main/res/layout/activity_task.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.TaskActivity">
+    tools:context=".ui.tasks.screens.TaskActivity">
 
     <LinearLayout
         android:id="@+id/linearLayout"
