+++ app/src/main/res/layout/fragment_compose_report_program.xml
@@ -7,7 +7,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:fitsSystemWindows="true"
     android:orientation="vertical"
-    tools:context=".mentortasks_ui.ComposeReportFragment">
+    tools:context=".ui.reports.screens.ComposeReportFragment">
     <androidx.appcompat.widget.Toolbar
         android:id="@+id/cps_toolbar"
         android:layout_width="match_parent"
