+++ app/src/main/res/layout/fragment_mentor_report_details.xml
@@ -4,7 +4,7 @@
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".mentortasks_ui.MentorReportDetailsFragment">
+    tools:context=".ui.reports.screens.MentorReportDetailsFragment">
 
     <TextView
         android:id="@+id/mrd_txt"
