--- app/src/main/res/layout/fragment_mentors_all_tasks.xml
@@ -1,84 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".mentortasks_ui.MentorsAllTasksFragment">
-
-    <include layout="@layout/include_toolbar" />
-
-    <LinearLayout
-        android:id="@+id/sps_case"
-        android:layout_width="@dimen/_300sdp"
-        android:layout_height="39sp"
-        android:orientation="horizontal"
-        android:layout_marginTop="@dimen/_4sdp"
-        android:layout_marginHorizontal="@dimen/margin_16"
-        android:background="@drawable/all_green_rounded_corner"
-        android:paddingStart="@dimen/_12ssp"
-        android:paddingEnd="@dimen/_8sdp"
-        android:paddingTop="@dimen/_6sdp"
-        android:paddingBottom="@dimen/_6sdp"
-        app:layout_constraintTop_toBottomOf="@+id/toolbar2"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent">
-
-        <TextView
-            android:id="@+id/txt_all"
-            android:layout_width="0dp"
-            android:layout_height="match_parent"
-            android:textColor="@color/MMS_pry_2"
-            android:gravity="center"
-            android:background="@drawable/all_bg"
-            android:layout_weight="0.5"
-            android:text="@string/all"
-            android:textSize="12sp" />
-
-        <TextView
-            android:id="@+id/txt_assigned"
-            android:layout_width="0dp"
-            android:layout_height="wrap_content"
-            android:textColor="@color/white"
-            android:gravity="center"
-            android:layout_weight="0.5"
-            android:text="@string/assigned"
-            android:textSize="12sp" />
-
-        <TextView
-            android:id="@+id/txt_completed"
-            android:layout_width="0dp"
-            android:layout_height="wrap_content"
-            android:textColor="@color/white"
-            android:gravity="center"
-            android:layout_weight="0.5"
-            android:text="@string/completed"
-            android:textSize="12sp" />
-
-        <TextView
-            android:id="@+id/txt_tasks"
-            android:layout_width="0dp"
-            android:layout_height="wrap_content"
-            android:textColor="@color/white"
-            android:gravity="center"
-            android:layout_weight="0.5"
-            android:text="@string/my_tasks"
-            android:textSize="12sp" />
-
-    </LinearLayout>
-
-    <androidx.recyclerview.widget.RecyclerView
-        android:id="@+id/listItems"
-        android:layout_width="match_parent"
-        android:layout_height="0dp"
-        android:layout_marginTop="8dp"
-        android:layout_marginBottom="8dp"
-        android:layout_marginHorizontal="@dimen/margin_8"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.0"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/sps_case"
-        app:layout_constraintVertical_bias="1.0"
-        tools:listitem="@layout/item_mentor_select_tasks" />
-</androidx.constraintlayout.widget.ConstraintLayout>
