+++ app/src/main/res/layout/fragment_mentors_report_select_tasks.xml
@@ -4,7 +4,7 @@
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".mentortasks_ui.MentorsReportSelectTasksFragment">
+    tools:context=".ui.reports.screens.MentorsReportSelectTasksFragment">
 
     <include layout="@layout/include_toolbar" />
