+++ app/src/main/res/layout/fragment_select_program.xml
@@ -8,7 +8,7 @@
     android:background="@color/white"
     android:fitsSystemWindows="true"
     android:orientation="vertical"
-    tools:context=".mentortasks_ui.SelectProgramFragment">
+    tools:context=".ui.reports.screens.SelectProgramFragment">
 
     <include layout="@layout/include_toolbar" />
