+++ app/src/main/res/layout/fragment_task_assigned.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".TaskAssignedFragment">
+    tools:context=".ui.tasks.screens.TaskAssignedFragment">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
