+++ app/src/main/res/navigation/nav_graph.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/composeReportFragment"
-        android:name="com.peculiaruc.alc_mmsystem_mentor.mentortasks_ui.ComposeReportFragment"
+        android:name="com.peculiaruc.alc_mmsystem_mentor.ui.reports.screens.ComposeReportFragment"
         android:label="fragment_compose_repot"
         tools:layout="@layout/fragment_compose_report_program" >
         <action
@@ -22,7 +22,7 @@
 
     <fragment
         android:id="@+id/mentorReportDetailsFragment"
-        android:name="com.peculiaruc.alc_mmsystem_mentor.mentortasks_ui.MentorReportDetailsFragment"
+        android:name="com.peculiaruc.alc_mmsystem_mentor.ui.reports.screens.MentorReportDetailsFragment"
         android:label="fragment_mentor_report_details"
         tools:layout="@layout/fragment_mentor_report_details" >
         <action
@@ -45,7 +45,7 @@
 
     <fragment
         android:id="@+id/mentorsReportFragment"
-        android:name="com.peculiaruc.alc_mmsystem_mentor.mentortasks_ui.MentorsReportFragment"
+        android:name="com.peculiaruc.alc_mmsystem_mentor.ui.reports.screens.MentorsReportFragment"
         android:label="fragment_mentors_report"
         tools:layout="@layout/fragment_mentors_report" >
 
@@ -61,7 +61,7 @@
 
     <fragment
         android:id="@+id/selectProgramFragment"
-        android:name="com.peculiaruc.alc_mmsystem_mentor.mentortasks_ui.SelectProgramFragment"
+        android:name="com.peculiaruc.alc_mmsystem_mentor.ui.reports.screens.SelectProgramFragment"
         android:label="fragment_select_program"
         tools:layout="@layout/fragment_select_program" >
         <action
@@ -76,13 +76,13 @@
 
     <fragment
         android:id="@+id/mentorReportSelectTasksFragment"
-        android:name="com.peculiaruc.alc_mmsystem_mentor.mentortasks_ui.MentorsReportSelectTasksFragment"
+        android:name="com.peculiaruc.alc_mmsystem_mentor.ui.reports.screens.MentorsReportSelectTasksFragment"
         android:label="fragment_mentors_select_tasks"
         tools:layout="@layout/fragment_mentors_report_select_tasks" />
 
     <fragment
         android:id="@+id/composeTaskFragment"
-        android:name="com.peculiaruc.alc_mmsystem_mentor.mentortasks_ui.ComposeReportTasksFragment"
+        android:name="com.peculiaruc.alc_mmsystem_mentor.ui.reports.screens.ComposeReportTasksFragment"
         android:label="fragment_compose_report_task"
         tools:layout="@layout/fragment_compose_report_tasks" >
         <action
@@ -97,13 +97,13 @@
 
     <fragment
         android:id="@+id/reportDownloadFragment"
-        android:name="com.peculiaruc.alc_mmsystem_mentor.mentortasks_ui.ReportDownloadFragment"
+        android:name="com.peculiaruc.alc_mmsystem_mentor.ui.reports.screens.ReportDownloadFragment"
         android:label="fragment_report_download"
         tools:layout="@layout/fragment_report_download"/>
 
     <fragment
         android:id="@+id/shareReportFragment"
-        android:name="com.peculiaruc.alc_mmsystem_mentor.mentortasks_ui.ShareReportFragment"
+        android:name="com.peculiaruc.alc_mmsystem_mentor.ui.reports.screens.ShareReportFragment"
         android:label="fragment_share_report"
         tools:layout="@layout/fragment_share_report">
         <action
@@ -126,7 +126,7 @@
 
     <fragment
         android:id="@+id/reportSubmittedFragment"
-        android:name="com.peculiaruc.alc_mmsystem_mentor.mentortasks_ui.ReportSubmittedFragment"
+        android:name="com.peculiaruc.alc_mmsystem_mentor.ui.reports.screens.ReportSubmittedFragment"
         android:label="fragment_report_submitted"
         tools:layout="@layout/fragment_report_submitted"/>
