+++ app/src/main/res/values/strings.xml
@@ -1,5 +1,5 @@
 <resources>
-    <string name="app_name">alc-MMSystem-mentor</string>
+    <string name="app_name">Mentor App</string>
     <string name="my_tasks">My Tasks</string>
     <string name="assign_task">Assign Task</string>
     <!-- TODO: Remove or change this placeholder text -->
